Fixed Random Joke/Made Button Joke
I overhauled the JS code for the random generated joke to account for both single and twopart jokes. I also chose to make sure the button for the programming joke also and made sure to have the id's for the rest of the buttons be made, also having comments to make it clear which section is happening first.