Title Change/Jokes
I added a id to the h1 one tag on the html in order to make it easier for me to call to it when the title changes based on the button pressed. I also added the rest of the joke types to the JavaScript for the rest of the buttons.